@@ -1,35 +1,27 @@
 package com.vincent.filepickersample;
 
+
+
 import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.View;
 import android.widget.TextView;
 
 import com.vincent.filepicker.Constant;
-import com.vincent.filepicker.Util;
 import com.vincent.filepicker.activity.AudioPickActivity;
+import com.vincent.filepicker.activity.BaseActivity;
 import com.vincent.filepicker.activity.ImagePickActivity;
 import com.vincent.filepicker.activity.NormalFilePickActivity;
 import com.vincent.filepicker.activity.VideoPickActivity;
-import com.vincent.filepicker.filter.FileFilter;
-import com.vincent.filepicker.filter.callback.FilterResultCallback;
 import com.vincent.filepicker.filter.entity.AudioFile;
-import com.vincent.filepicker.filter.entity.BaseFile;
-import com.vincent.filepicker.filter.entity.Directory;
 import com.vincent.filepicker.filter.entity.ImageFile;
 import com.vincent.filepicker.filter.entity.NormalFile;
 import com.vincent.filepicker.filter.entity.VideoFile;
 
 import java.util.ArrayList;
-import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
-import static com.vincent.filepicker.activity.AudioPickActivity.IS_NEED_RECORDER;
-import static com.vincent.filepicker.activity.BaseActivity.IS_NEED_FOLDER_LIST;
-import static com.vincent.filepicker.activity.ImagePickActivity.IS_NEED_CAMERA;
+
 
 public class MainActivity extends AppCompatActivity implements View.OnClickListener{
     private TextView mTvResult;
@@ -48,29 +40,29 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         switch (id) {
             case R.id.btn_pick_image:
                 Intent intent1 = new Intent(this, ImagePickActivity.class);
-                intent1.putExtra(IS_NEED_CAMERA, true);
+                intent1.putExtra(ImagePickActivity.IS_NEED_CAMERA, true);
                 intent1.putExtra(Constant.MAX_NUMBER, 9);
-                intent1.putExtra(IS_NEED_FOLDER_LIST, true);
+                intent1.putExtra(BaseActivity.IS_NEED_FOLDER_LIST, true);
                 startActivityForResult(intent1, Constant.REQUEST_CODE_PICK_IMAGE);
                 break;
             case R.id.btn_pick_video:
                 Intent intent2 = new Intent(this, VideoPickActivity.class);
-                intent2.putExtra(IS_NEED_CAMERA, true);
+                intent2.putExtra(VideoPickActivity.IS_NEED_CAMERA, true);
                 intent2.putExtra(Constant.MAX_NUMBER, 9);
-                intent2.putExtra(IS_NEED_FOLDER_LIST, true);
+                intent2.putExtra(BaseActivity.IS_NEED_FOLDER_LIST, true);
                 startActivityForResult(intent2, Constant.REQUEST_CODE_PICK_VIDEO);
                 break;
             case R.id.btn_pick_audio:
                 Intent intent3 = new Intent(this, AudioPickActivity.class);
-                intent3.putExtra(IS_NEED_RECORDER, true);
+                intent3.putExtra(AudioPickActivity.IS_NEED_RECORDER, true);
                 intent3.putExtra(Constant.MAX_NUMBER, 9);
-                intent3.putExtra(IS_NEED_FOLDER_LIST, true);
+                intent3.putExtra(BaseActivity.IS_NEED_FOLDER_LIST, true);
                 startActivityForResult(intent3, Constant.REQUEST_CODE_PICK_AUDIO);
                 break;
             case R.id.btn_pick_file:
                 Intent intent4 = new Intent(this, NormalFilePickActivity.class);
                 intent4.putExtra(Constant.MAX_NUMBER, 9);
-                intent4.putExtra(IS_NEED_FOLDER_LIST, true);
+                intent4.putExtra(BaseActivity.IS_NEED_FOLDER_LIST, true);
                 intent4.putExtra(NormalFilePickActivity.SUFFIX,
                         new String[] {"xlsx", "xls", "doc", "dOcX", "ppt", ".pptx", "pdf"});
                 startActivityForResult(intent4, Constant.REQUEST_CODE_PICK_FILE);
@@ -80,10 +72,11 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
 
     @Override
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
         switch (requestCode) {
             case Constant.REQUEST_CODE_PICK_IMAGE:
                 if (resultCode == RESULT_OK) {
-                    ArrayList<ImageFile> list = data.getParcelableArrayListExtra(Constant.RESULT_PICK_IMAGE);
+                    ArrayList< ImageFile > list = data.getParcelableArrayListExtra(Constant.RESULT_PICK_IMAGE);
                     StringBuilder builder = new StringBuilder();
                     for (ImageFile file : list) {
                         String path = file.getPath();
@@ -94,7 +87,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
                 break;
             case Constant.REQUEST_CODE_PICK_VIDEO:
                 if (resultCode == RESULT_OK) {
-                    ArrayList<VideoFile> list = data.getParcelableArrayListExtra(Constant.RESULT_PICK_VIDEO);
+                    ArrayList< VideoFile > list = data.getParcelableArrayListExtra(Constant.RESULT_PICK_VIDEO);
                     StringBuilder builder = new StringBuilder();
                     for (VideoFile file : list) {
                         String path = file.getPath();
@@ -105,7 +98,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
                 break;
             case Constant.REQUEST_CODE_PICK_AUDIO:
                 if (resultCode == RESULT_OK) {
-                    ArrayList<AudioFile> list = data.getParcelableArrayListExtra(Constant.RESULT_PICK_AUDIO);
+                    ArrayList< AudioFile > list = data.getParcelableArrayListExtra(Constant.RESULT_PICK_AUDIO);
                     StringBuilder builder = new StringBuilder();
                     for (AudioFile file : list) {
                         String path = file.getPath();
@@ -116,7 +109,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
                 break;
             case Constant.REQUEST_CODE_PICK_FILE:
                 if (resultCode == RESULT_OK) {
-                    ArrayList<NormalFile> list = data.getParcelableArrayListExtra(Constant.RESULT_PICK_FILE);
+                    ArrayList< NormalFile > list = data.getParcelableArrayListExtra(Constant.RESULT_PICK_FILE);
                     StringBuilder builder = new StringBuilder();
                     for (NormalFile file : list) {
                         String path = file.getPath();
