@@ -1,7 +1,7 @@
 package com.vincent.filepicker.adapter;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -9,10 +9,8 @@ import android.widget.TextView;
 
 import com.vincent.filepicker.R;
 import com.vincent.filepicker.filter.entity.Directory;
-import com.vincent.filepicker.filter.entity.ImageFile;
 
 import java.util.ArrayList;
-import java.util.List;
 
 /**
  * Created by Vincent Woo
