@@ -3,9 +3,7 @@ package com.vincent.filepicker.filter.loader;
 import android.content.Context;
 import android.net.Uri;
 import android.provider.MediaStore;
-import android.support.v4.content.CursorLoader;
-
-import static android.provider.MediaStore.MediaColumns.MIME_TYPE;
+import androidx.loader.content.CursorLoader;
 
 /**
  * Created by Vincent Woo
